Sample-Code-Snippets/Python/Azure Function Example/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aiohttp
azure-functions-durable
azure-eventhub
azure-eventhub-checkpointstoreblob-aio